requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
multidict==5.1.0
mysql-connector-python==8.0.25
Pillow==8.2.0
protobuf==3.17.3
pyTelegramBotAPI==3.7.9
requests==2.25.1
six==1.16.0
telepot==12.7
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3
